Перешел на WorkManager

+++ app/src/main/java/com/luigivampa92/yms/fintracker/model/Record.kt
@@ -15,6 +15,6 @@ data class Record(
         @ColumnInfo() val currency: String,
         @ColumnInfo() val wallet_id: String,
         @ColumnInfo() val date: String,
-        @ColumnInfo() val pending_time: Long = 0L,
+        @ColumnInfo() val pending_time: Int = 0,
         @ColumnInfo() val repeatable: Boolean = false
 )
